pixma: A few changes for MX492.
Will shortly refacture image_size, line_size and other scan parameters
to be able to differentiate between our interactions with the scanner
and the user, because these value might be different such as in the
case of software_lineart where the user is getting lineart, but the
scanner is giving us grey. The difference in the values of these
critical parameters is what is causing most of the issues.