pixma: convert line_size, image_size to frontend and backend equivalents
These are computed values for the coming scan and they are prefixed with
be_ and fe_ for frontend and backend respectively.
"frontend" here means what the user requests, and "backend" means what
we are asking of the scanner. This is mostly the same, but when doing
lineart, the backend is actually doing grey, since lineart is not
directly supported.
Many of the bugs here are related to the incorrect use of the fe_ values
when communicating with the scanner during the scan and it completely
messes up our processing of the raw scan data. The separation makes it
clearer what the parameters refer to.

This concludes the support for the MX492 which is now fully functional.